+++ src/DataObjects/CustomerData.php
@@ -4,8 +4,9 @@ declare(strict_types=1);
 
 namespace Subster\PhpSdk\DataObjects;
 
-use Carbon\CarbonImmutable;
+use DateTimeImmutable;
 use Subster\PhpSdk\Concerns\Data;
+use Subster\PhpSdk\Support\DateTimeNormalizer;
 
 class CustomerData extends Data
 {
@@ -13,8 +14,8 @@ class CustomerData extends Data
         public readonly string $id,
         public readonly ?string $name,
         public readonly string $email,
-        public readonly CarbonImmutable $created_at,
-        public readonly CarbonImmutable $updated_at,
+        public readonly DateTimeImmutable $created_at,
+        public readonly DateTimeImmutable $updated_at,
     ) {}
 
     public static function fromSaloon(array $response): self
@@ -23,8 +24,8 @@ class CustomerData extends Data
             id: strval($response['id']),
             name: isset($response['name']) ? strval($response['name']) : null,
             email: strval($response['email']),
-            created_at: CarbonImmutable::parse($response['created_at']),
-            updated_at: CarbonImmutable::parse($response['updated_at']),
+            created_at: DateTimeNormalizer::parse($response['created_at']),
+            updated_at: DateTimeNormalizer::parse($response['updated_at']),
         );
     }
 }